chore: re-introduce separate types for ifEventIsMissing
