+++ pkg/eventsourcingdb/ifeventismissingduringobserve/documentation.go
@@ -1,0 +1,2 @@
+// Package ifeventismissingduringobserve provides a single enum for the fromLatestEvent option.
+package ifeventismissingduringobserve
