+++ pkg/eventsourcingdb/ifeventismissingduringobserve/if_event_is_missing.go
@@ -1,9 +1,8 @@
-package eventsourcingdb
+package ifeventismissingduringobserve
 
 type IfEventIsMissing string
 
 const (
-	ReadNothing    IfEventIsMissing = "read-nothing"
 	ReadEverything IfEventIsMissing = "read-everything"
 	WaitForEvent   IfEventIsMissing = "wait-for-event"
 )
