+++ pkg/eventsourcingdb/ifeventismissingduringread/documentation.go
@@ -1,0 +1,2 @@
+// Package ifeventismissingduringread provides a single enum for the fromLatestEvent option.
+package ifeventismissingduringread
