+++ pkg/eventsourcingdb/ifeventismissingduringread/if_event_is_missing.go
@@ -1,0 +1,8 @@
+package ifeventismissingduringread
+
+type IfEventIsMissing string
+
+const (
+	ReadEverything IfEventIsMissing = "read-everything"
+	ReadNothing    IfEventIsMissing = "read-nothing"
+)
