+++ pkg/eventsourcingdb/read_events_options.go
@@ -2,8 +2,8 @@ package eventsourcingdb
 
 import (
 	"errors"
-	"fmt"
 	"github.com/thenativeweb/eventsourcingdb-client-golang/pkg/eventsourcingdb/event"
+	"github.com/thenativeweb/eventsourcingdb-client-golang/pkg/eventsourcingdb/ifeventismissingduringread"
 	"strconv"
 )
 
@@ -22,9 +22,9 @@ func ReadNonRecursively() ReadRecursivelyOption {
 }
 
 type readFromLatestEvent struct {
-	Subject          string           `json:"subject"`
-	Type             string           `json:"type"`
-	IfEventIsMissing IfEventIsMissing `json:"ifEventIsMissing"`
+	Subject          string                                      `json:"subject"`
+	Type             string                                      `json:"type"`
+	IfEventIsMissing ifeventismissingduringread.IfEventIsMissing `json:"ifEventIsMissing"`
 }
 
 type readEventsOptions struct {
@@ -38,17 +38,6 @@ type readEventsOptions struct {
 type ReadEventsOption struct {
 	apply func(options *readEventsOptions) error
 	name  string
-}
-
-func validateIfEventIsMissingForReadEvents(value IfEventIsMissing) error {
-	switch value {
-	case ReadEverything:
-		fallthrough
-	case ReadNothing:
-		return nil
-	default:
-		return fmt.Errorf("%q is not a valid value for ifEventIsMissing, it must be either %q or %q", value, ReadEverything, ReadNothing)
-	}
 }
 
 func ReadChronologically() ReadEventsOption {
@@ -117,7 +106,7 @@ func ReadUntilUpperBoundID(upperBoundID string) ReadEventsOption {
 	}
 }
 
-func ReadFromLatestEvent(subject, eventType string, ifEventIsMissing IfEventIsMissing) ReadEventsOption {
+func ReadFromLatestEvent(subject, eventType string, ifEventIsMissing ifeventismissingduringread.IfEventIsMissing) ReadEventsOption {
 	return ReadEventsOption{
 		apply: func(options *readEventsOptions) error {
 			if options.LowerBoundID != nil {
@@ -127,9 +116,6 @@ func ReadFromLatestEvent(subject, eventType string, ifEventIsMissing IfEventIsMi
 				return err
 			}
 			if err := event.ValidateType(eventType); err != nil {
-				return err
-			}
-			if err := validateIfEventIsMissingForReadEvents(ifEventIsMissing); err != nil {
 				return err
 			}
